Delete Work.tsx and switch to dynamic routing
